[Refactoring] : 도메인(로그인/회원가입) 분리
