+++ src/features/auth/api/index.ts
@@ -1,1 +1,2 @@
 export * from './auth.api';
+export * from './email.api';
