+++ src/features/auth/hooks/login/useLoginMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { authService } from '../api/auth.api';
+import { authService } from '../../api/auth.api';
 import { tokenManager } from '@/shared/libs/auth/tokenManager';
-import { LoginResponse } from '../types/auth';
+import { LoginResponse } from '../../types/auth';
 import { handleAPIError } from '@/shared/utils/errorHandler';
 
 export const useLoginMutation = () => {
