+++ src/features/auth/ui/login/LoginForm.tsx
@@ -1,5 +1,5 @@
 import { AiOutlineEye, AiOutlineEyeInvisible } from 'react-icons/ai';
-import { useLoginForm } from '../hooks';
+import { useLoginForm } from '../../hooks';
 
 export const LoginForm = () => {
   const {
