+++ src/features/auth/ui/signup/index.ts
@@ -1,0 +1,1 @@
+export * from './SignupForm';
